Функции связанные с Render перенесены в /utils/render.js
