+++ src/main.js
@@ -5,10 +5,10 @@ import RankView from './view/user-rating/create-user-rank.js';
 import FilmListView from './view/films.js';
 import FilmSummaryStatisticView from './view/statistic-summary.js';
 import {generateFilms} from './mock/generate-films.js';
-import {renderDOMStrings, RenderPosition} from './view/utils.js';
+import {renderDOMStrings, RenderPosition} from './view/utils/render.js';
 import {countedStatistics, statistics} from './view/statistic/count-statistics.js';
 
-const FILMS_COUNT = 16;
+const FILMS_COUNT = 1000;
 const header = document.querySelector('.header');
 const main = document.querySelector('.main');
 const statisticsElement = document.querySelector('.footer__statistics');
